Added APPS1 and APPS2 10% difference fault
- Changed FP_sensor_diff to type float from type uint
- I will test this

+++ VCU/Phantom/main.c
@@ -143,7 +143,7 @@ uint16 FP_sensor_1_max = 4095; // 12-bit ADC
 uint16 FP_sensor_2_max = 4095; // 12-bit ADC
 uint16 FP_sensor_1_percentage;
 uint16 FP_sensor_2_percentage;
-uint16 FP_sensor_diff;
+float FP_sensor_diff;
